Refactoring) 코드리뷰 반영 및 RestaurantDetail 컴포넌트명 변경
