+++ src/About.test.jsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 
 import AboutPage from './AboutPage';
 
-test('About', () => {
+test('AboutPage', () => {
   const { queryByText } = render((<AboutPage />));
 
   expect(queryByText(/About 페이지 입니다./)).not.toBeNull();
