+++ src/App.jsx
@@ -3,10 +3,10 @@ import React from 'react';
 import { Switch, Route, Link } from 'react-router-dom';
 
 import HomePage from './HomePage';
-import RestaurantsPage from './RestaurantsPage';
+import RestaurantsPage from './restaurants/RestaurantsPage';
 import AboutPage from './AboutPage';
 import NotFoundPage from './NotFoundPage';
-import RestaurantDetailPage from './RestaurantDetailPage';
+import DetailPage from './restaurants/DetailPage';
 
 export default function App() {
   return (
@@ -17,8 +17,8 @@ export default function App() {
       <Switch>
         <Route exact path="/" component={HomePage} />
         <Route exact path="/restaurants" component={RestaurantsPage} />
+        <Route path="/restaurants/:id" component={DetailPage} />
         <Route path="/about" component={AboutPage} />
-        <Route path="/restaurants/:id" component={RestaurantDetailPage} />
         <Route component={NotFoundPage} />
       </Switch>
     </div>
