+++ src/restaurants/CategoriesContainer.test.jsx
