+++ src/restaurants/DetailPage.jsx
@@ -1,16 +1,18 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import { loadRestaurant } from './actions';
+import { loadRestaurant } from '../actions';
 
 import MenuItems from './MenuItems';
 
-export default function RestaurantDetailPage({ match }) {
+export default function DetailPage({ match }) {
   const dispatch = useDispatch();
   const { id } = match.params;
 
-  dispatch(loadRestaurant(id));
+  useEffect(() => {
+    dispatch(loadRestaurant(id));
+  });
 
   const { name, address, menuItems } = useSelector((state) => state.restaurant);
