+++ src/restaurants/RestaurantsContainer.jsx
@@ -4,7 +4,7 @@ import { useSelector } from 'react-redux';
 
 import { Link } from 'react-router-dom';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function RestaurantsContainer() {
   const restaurants = useSelector(get('restaurants'));
@@ -13,7 +13,7 @@ export default function RestaurantsContainer() {
     <ul>
       {restaurants.map((restaurant) => (
         <li key={restaurant.id}>
-          <Link to={`/restaurants/${restaurant.id}`}>
+          <Link to={`restaurants/${restaurant.id}`}>
             {restaurant.name}
           </Link>
         </li>
